Rename file and function names and add docstring
