+++ mythril/analysis/analysis_utils.py
@@ -1,9 +1,16 @@
 import re
+from typing import List
 from z3 import *
 from mythril.laser.ethereum.transaction import ContractCreationTransaction
+from mythril.laser.ethereum.state import GlobalState
 
 
-def get_non_creator_constraints(state, node):
+def get_non_creator_constraints(state: GlobalState) -> (List, bool):
+    """
+    Get constraints which say that the caller isn't the creator of the contract
+    :param state: The state
+    :return: tuple of (constraints, bool) where the bool says whether the caller is constrained or not
+    """
     not_creator_constraints = []
     creator = None
     if isinstance(
@@ -23,12 +30,15 @@ def get_non_creator_constraints(state, node):
             not_creator_constraints.append(
                 Not(Extract(159, 0, transaction.caller) == 0)
             )
-        if not _check_changeable_constraints(node.constraints):
+        if not has_caller_check_constraint(state.mstate.constraints):
             return [], True
     return not_creator_constraints, False
 
 
-def _check_changeable_constraints(constraints):
+def has_caller_check_constraint(constraints: List) -> bool:
+    """
+    Checks whether the caller is constrained to a value or not
+    """
     for constraint in constraints:
         if re.search(r"caller", str(constraint)) and re.search(
             r"[0-9]{20}", str(constraint)
