+++ mythril/analysis/modules/suicide.py
@@ -1,5 +1,5 @@
 from mythril.analysis import solver
-from mythril.analysis.get_custom_constraints import get_non_creator_constraints
+from mythril.analysis.analysis_utils import get_non_creator_constraints
 from mythril.analysis.ops import *
 from mythril.analysis.report import Issue
 from mythril.analysis.swc_data import UNPROTECTED_SELFDESTRUCT
